+++ src/lib.rs
@@ -5,7 +5,7 @@ mod error;
 mod signing;
 
 pub use admin_websocket::{AdminWebsocket, AuthorizeSigningCredentialsPayload, EnableAppResponse};
-pub use app_agent_websocket::{AppAgentWebsocket, ZomeCallTarget};
+pub use app_agent_websocket::{AppWebsocket, ZomeCallTarget};
 pub use error::{ConductorApiError, ConductorApiResult};
 pub use holochain_conductor_api::{
     AdminRequest, AdminResponse, AppAuthenticationRequest, AppAuthenticationToken,